Revert "Bug 1089718 - [music] Make the metadata parser sane"
