+++ apps/music/js/utils.js
@@ -1,5 +1,6 @@
-/* exported formatTime, createListElement */
-/* global Normalizer, AlbumArt */
+/* exported formatTime, simpleHash, generateDefaultThumbnailURL,
+            getAlbumArtBlob, getAlbumArtBlob, createListElement */
+/* global getThumbnailURL, Normalizer */
 'use strict';
 
 function formatTime(secs) {
@@ -26,6 +27,72 @@ function formatTime(secs) {
   }
 
   return formatedTime;
+}
+
+// The Javascript implementation of Java’s String.hashCode() method.
+function simpleHash(str) {
+  var hash = 0;
+  if (str.length === 0) {
+    return hash;
+  }
+
+  for (var i = 0; i < str.length; i++) {
+    var c = str.charCodeAt(i);
+    hash = ((hash << 5) - hash) + c;
+    hash = hash & hash; // Convert to 32-bit integer
+  }
+
+  return hash;
+}
+
+function generateDefaultThumbnailURL(metadata) {
+  // If metadata does not contain both album and artist, then use title instead.
+  var infoForHash = (!metadata.album && !metadata.artist) ?
+    metadata.title : metadata.album + metadata.artist;
+  var hashedNumber = (Math.abs(simpleHash(infoForHash)) % 10) + 1;
+
+  return '/style/images/AlbumArt' + hashedNumber + '_small.png';
+}
+
+// Fetch the cover art for a given file and return it as a Blob. If there's no
+// embedded cover art, grab the appropriate placeholder image instead. XXX: This
+// function is a bit convoluted, and we could stand to simplify things here,
+// e.g. by merging placeholder art with embedded art and/or storing cover art
+// on the sdcard.
+function getAlbumArtBlob(fileinfo, callback) {
+  var getBlob = function(url, callback) {
+    var xhr = new XMLHttpRequest();
+    xhr.open('GET', url, true);
+    xhr.responseType = 'blob';
+
+    xhr.onload = function() {
+      callback(null, xhr.response);
+    };
+    // I don't think onerror usually gets called, but let's play it safe.
+    xhr.onerror = function() {
+      callback('error', null);
+    };
+
+    // Bad local URLs throw in send() for some reason.
+    try {
+      xhr.send();
+    } catch (e) {
+      callback(e, null);
+    }
+  };
+
+  if ('picture' in fileinfo.metadata) {
+    getThumbnailURL(fileinfo, function(url) {
+      if (!url) {
+        return callback(null, null);
+      }
+      getBlob(url, callback);
+    });
+  }
+  else {
+    var url = generateDefaultThumbnailURL(fileinfo.metadata);
+    getBlob(url, callback);
+  }
 }
 
 // In Music, visually we have three styles of list
@@ -93,9 +160,12 @@ function createListElement(option, data, index, highlight) {
       // Use background image instead of creating img elements can reduce
       // the amount of total elements in the DOM tree, it can save memory
       // and gecko can render the elements faster as well.
-      AlbumArt.getCoverURL(data).then(function(url) {
+      var setBackground = function(url) {
+        url = url || generateDefaultThumbnailURL(data.metadata);
         li.style.backgroundImage = 'url(' + url + ')';
-      });
+      };
+
+      getThumbnailURL(data, setBackground);
 
       if (option === 'artist') {
         artistSpan = document.createElement('span');
