+++ apps/music/test/unit/mp4tags_test.js
@@ -1,28 +1,12 @@
-/* global parseMetadata, MockLazyLoader, MockGetDeviceStorage */
+/* global parseMetadata */
 'use strict';
 
-require('/test/unit/metadata/utils.js');
-require('/js/metadata/mp4.js');
 
-suite('mp4 tags', function() {
-  var RealLazyLoader, RealGetDeviceStorage;
+require('/test/unit/metadata_utils.js');
 
-  setup(function(done) {
+suite('m4a tags', function() {
+  setup(function() {
     this.timeout(1000);
-    RealLazyLoader = window.LazyLoader;
-    window.LazyLoader = MockLazyLoader;
-
-    RealGetDeviceStorage = navigator.getDeviceStorage;
-    navigator.getDeviceStorage = MockGetDeviceStorage;
-
-    require('/js/metadata_scripts.js', function() {
-      done();
-    });
-  });
-
-  teardown(function() {
-    window.LazyLoader = RealLazyLoader;
-    navigator.getDeviceStorage = RealGetDeviceStorage;
   });
 
   test('m4a tags', function(done) {
